De-Windows-ify and generalize sys.path hacking
Fix indentation

Modified __import__('') test; in CPython this raises a ValueError, as
it now does in JPython.

In JPython now, Python keywords used as attributes works without `_'
appending, but the test wasn't updated for this.  Also, if
python.deprecated.keywordMangling is true in the registry then
underscore mangling does still work.  Fix the assert test of this
feature.


git-svn-id: https://jython.svn.sourceforge.net/svnroot/jython/trunk/jython@487 79b11f02-f210-0410-99aa-fb189b3d1ac4