Several changes, file-wide.
All values dug out of the 9-tuple are now checked for validity of
range when used.  ValueError is raised if out-of-range.

strftime(): Java's notion of "locale's time representation" and
"locale's date representation" are different than C's strftime() for
the "C" locale.  Java's default locale around here is en_US, and while
I can create and use a "C" locale, it isn't one of the listed
available locales, so Java either silently ignores it or equates it to
en_US.

This means that %x, %X, and %Z have different representations than
expected out of C's strftime().  %x and %Z are non-standard, so I'm
leaving them as-is.  I am hardcoding %X to HH:mm:SS so that the
majority of test_strftime.py tests pass without complaint.


git-svn-id: https://jython.svn.sourceforge.net/svnroot/jython/trunk/jython@508 79b11f02-f210-0410-99aa-fb189b3d1ac4