Patch provided by Drew Morrissey <drew.morrissey@voiceint.com> and approved by JimH, this reaps dead threads from the thread state cache so as to avoid memory leaks when many threads are used. Specifically,
additionCounter: new variable to count the number of new threads added
to the cache.

MAX_ADDITIONS: threshold beyond which dead thread reaping loop is
called.

cleanupThreadTable(): loop to reap all !isAlive() threads.

getThreadState(PySystemState): every time a new thread is added to the
cache, increment the additionCounter.  When this is > MAX_ADDITIONS,
run the cleanup loop and reset the counter.


git-svn-id: https://jython.svn.sourceforge.net/svnroot/jython/trunk/jython@829 79b11f02-f210-0410-99aa-fb189b3d1ac4