Patch for PR#195 (and other related bugs)
The problem was caused when an interface appeared multiple times in
the inheritance hierarchy through a diamond path.  E.g.

             Model (a Python class)
               |
  +------------+-------------+
  |                          |
  |extends                   |implements
  |                          |
  V                          V
AbstractListModel       ComboBoxModel
  |                          |
  |implements                |extends
  |                          |
  +------------+-------------+
               |
           ListModel

We were not keeping track of all methods seen throughout the entire
hierarchy.  This patch fixes this by removing the addMethods(Class)
method and forcing everything to use addMethods(Class, Hashtable).


git-svn-id: https://jython.svn.sourceforge.net/svnroot/jython/trunk/jython@836 79b11f02-f210-0410-99aa-fb189b3d1ac4