com.oroinc.text.regex => org.apache.oro
some minor formatting nits

getindex(): If we get an exception converting s to an integer, then
test to see if the string is a valid Python identifier.  If not, raise
a re.error instead of an IndexError.  This is for CPython
compatibility, which requires the group names to be integers or Python
identifiers.

isdigit(), isident(), isname(): support functions for the "valid
Python identifier" test.

fixPattern(): If a group name is specified, make sure it's a valid
Python identifier.  Otherwise throw a re.error.

findall(): Doh!  Support re.findall() after all. (I actually added
this a while ago and forgot :)


git-svn-id: https://jython.svn.sourceforge.net/svnroot/jython/trunk/jython@1138 79b11f02-f210-0410-99aa-fb189b3d1ac4