List comprehension support.
The grammer used is slightly different from CPython's. Where CPyton use

   //list_iter: list_for | list_if
   //list_for: 'for' exprlist 'in' testlist [list_iter]
   //list_if: 'if' test [list_iter]

this patch uses

   //list_iter: [list_for | list_if]
   //list_for: 'for' exprlist 'in' testlist list_iter
   //list_if: 'if' test list_iter

This mean that the chain of statements inside a list comprehension
expression always ends with a empty (without any children) list_iter
node. This empty node is used to trigger the call to the append method.

In CodeCompiler the list is kept on the stack and the append method
is stored in a register.

In jythonc all the code required to calculate the list is placed
in a separate private method called "__listcomprehension$1".
The list comprehension expression is replaced with a call to this
private method.

In both code generators the element expression node and the
"[].append" method stored on a stack when a list comprehension
is recognized and used when the empty list_iter node is reached.
The actual if/for statement are generated by the existing if/for
code.


git-svn-id: https://jython.svn.sourceforge.net/svnroot/jython/trunk/jython@1160 79b11f02-f210-0410-99aa-fb189b3d1ac4