Commit of akuchling's fix for bug #1599004
Things fixed from his message:

* SHA and MD5 objects lack a .digest_size attribute.
* Successive calls to digest() return different values.
* For digest(), the byte array is converted to a string using the default encoding, which produces weird results if it's UTF-8. It should hardwire using ISO-8859-1, because we want a 1-to-1 mapping of characters.

I also added a to_bytes method to PyString that handles getting the bytes to
add to a digest.  It takes a String an returns a byte array and will throw an
exception if there are for values in the string above 255.  Using
getBytes("ISO-8859-1") just skips over illegal values.



git-svn-id: https://jython.svn.sourceforge.net/svnroot/jython/trunk/jython@3038 79b11f02-f210-0410-99aa-fb189b3d1ac4