Merging the astwrite branch into trunk.  While I was testing, I noticed some problems stemming from the package names and the "Type" at the end of some of the asdl_antlr.py generated code -- so I fixed that up as well.
This checkin changes all of the access on ast nodes to be getters and setters
with Lists instead of arrays.  The ast nodes also now descend from PyObject
instead of Antlr's CommonTree.  There are new adapter classes that handle the
back and forth of ast nodes from Java to Python.

Merged revisions 5586-5592,5594-5595,5597,5612,5628,5633-5634,5636-5637,5641-5643,5660,5672,5677-5679,5683,5685-5686 via svnmerge from 
https://jython.svn.sourceforge.net/svnroot/jython/branches/astwrite



git-svn-id: https://jython.svn.sourceforge.net/svnroot/jython/trunk/jython@5695 79b11f02-f210-0410-99aa-fb189b3d1ac4