test302 - Move to test_pep263_jy
This adds a check for the encoding declared by the source file actually
matching up with what we were getting out of it, and switches the default
encoding to ascii to follow CPython. CPython assumes utf-8 in some single
compilation contexts, and this uses the same parser algorithm everywhere, so I
made a small change to test_doctest since Jython is throwing a SyntaxError on
utf-8 strings to a single compilation.



git-svn-id: https://jython.svn.sourceforge.net/svnroot/jython/trunk/jython@5847 79b11f02-f210-0410-99aa-fb189b3d1ac4