Add a jython initializer service.
If META-INF/services/org.python.core.JythonInitializer is on the classpath, the class named in that
file will be instantiated and used in Jython's initialization.  This is useful when Jython is
initialized by a library outside of your control, but some customization still needs to be done to
Jython's environment.  I promise not to add a JythonFactory or a JythonFactoryFactory.



git-svn-id: https://jython.svn.sourceforge.net/svnroot/jython/trunk/jython@5955 79b11f02-f210-0410-99aa-fb189b3d1ac4